// src/frontend/enums/route.enum.ts
export enum Route {
  HOME = '/',
  MATCH = '/match',
  MY_PARTNERS = '/my-partners',
  CHAT = '/chat',
  SESSIONS = '/sessions',
  POKEMON = '/pokemon',
  CATEGORIES = '/categories',
  PRODUCTS = '/products',
  LOGIN = '/login',
  SIGNUP = '/signup',
  ONBOARDING = '/onboarding',
  PROFILE = '/profile',
}

export enum RouteSlug {
  HOME = 'home',
  MATCH = 'match',
  POKEMON = 'pokemon',
  CATEGORIES = 'categories',
  PRODUCTS = 'products',
  LOGIN = 'login',
  SIGNUP = 'signup',
  ONBOARDING = 'onboarding',
}